+++ OCT2SysDisease/radiomics/network_3D_INL_RPE_indexSpace/generate3D100Radiomics.py
@@ -71,7 +71,7 @@ def main():
 
         nFeatures = 0
         sortedFeatureKeys = sorted(featureVector.keys())  # make sure the output value in dictionary order.
-        radiomicsArray = np.zeros((1, K), dtype=np.float32)
+        radiomicsArray = np.zeros((1, K), dtype=float)
         for featureName in sortedFeatureKeys:
             if "original_" == featureName[0:9]:
                 # print(f"{featureName}:{featureVector[featureName]}")
