+++ OCTMultiSurfaces/MICCAI2020Model/OCTDataSet.py
@@ -16,12 +16,12 @@ class OCTDataSet(data.Dataset):
         self.m_sigma = sigma
 
         # image uses float32
-        images = torch.from_numpy(np.load(imagesPath).astype(np.float32)).to(self.m_device, dtype=torch.float)  # slice, H, W
+        images = torch.from_numpy(np.load(imagesPath).astype(float)).to(self.m_device, dtype=torch.float)  # slice, H, W
         # normalize images for each slice
         std,mean = torch.std_mean(images, dim=(1,2))
         self.m_images = TF.Normalize(mean, std)(images)
 
-        self.m_labels = torch.from_numpy(np.load(labelPath).astype(np.float32)).to(self.m_device, dtype=torch.float)  # slice, num_surface, W
+        self.m_labels = torch.from_numpy(np.load(labelPath).astype(float)).to(self.m_device, dtype=torch.float)  # slice, num_surface, W
         with open(IDPath) as f:
             self.m_IDs = json.load(f)
         self.m_transform = transform
