+++ OCTMultiSurfaces/SoftSepar3Unet/afterLambda_UnconstrainedOpt.py
@@ -66,7 +66,7 @@ def main():
 
     # get Sigma2, Mu, R, r, G
     sigma2 = np.load(sigma2Path) # size: 3009x3x361
-    mu = np.load(muPath).astype(np.float32)
+    mu = np.load(muPath).astype(float)
     g  = np.load(gPath)
     r  = np.load(rPath)    # size:3009x2x361
     riftGT = np.load(riftGTPath)
